refactor(sdk/rust): update project layout
